refactor: table row layout
