+++ src/components/main/Liquidations/LiquidationsMetrics.tsx
@@ -7,8 +7,6 @@ export default function LiquidationsMetrics() {
   const { data: liquidityOverviewData, isLoading: isLiquidityOverviewDataLoading } =
     useOverviewData()
 
-  console.log(liquidityOverviewData, 'liquidityOverviewData')
-
   const badDebt = liquidityOverviewData?.data[0].bad_debt
   const walletsForLiquidation = liquidityOverviewData?.data[0].wallets_for_liquidation
   const walletsAtRisk = liquidityOverviewData?.data[0].wallets_at_risk
